LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-select-element/customizable-select/select-home-end-pagedown-pageup-detailed.optional-expected.txt


Option #1
Option #2
Option #3
Option #4
Option #5
Option #6
Option #7
Option #8
Option #9
Option #10
Option #11
Option #12
Option #13
Option #14
Option #15
Option #16
Option #17
Option #18
Option #19
Option #20
Option #21
Option #22
Option #23
Option #24
Option #25
Option #26
Option #27
Option #28
Option #29
Option #30
Option #31
Option #32
Option #33
Option #34
Option #35
Option #36
Option #37
Option #38
Option #39
Option #40
Option #41
Option #42
Option #43
Option #44
Option #45
Option #46
Option #47
Option #48
Option #49
Option #50
Option #51
Option #52
Option #53
Option #54
Option #55
Option #56
Option #57
Option #58
Option #59
Option #60
Option #61
Option #62
Option #63
Option #64
Option #65
Option #66
Option #67
Option #68
Option #69
Option #70
Option #71
Option #72
Option #73
Option #74
Option #75
Option #76
Option #77
Option #78
Option #79
Option #80
Option #81
Option #82
Option #83
Option #84
Option #85
Option #86
Option #87
Option #88
Option #89
Option #90
Option #91
Option #92
Option #93
Option #94
Option #95
Option #96
Option #97
Option #98
Option #99
Option #100
Option #101
Option #102
Option #103
Option #104
Option #105
Option #106
Option #107
Option #108
Option #109
Option #110
Option #111
Option #112
Option #113
Option #114
Option #115
Option #116
Option #117
Option #118
Option #119
Option #120
Option #121
Option #122
Option #123
Option #124
Option #125
Option #126
Option #127
Option #128
Option #129
Option #130
Option #131
Option #132
Option #133
Option #134
Option #135
Option #136
Option #137
Option #138
Option #139
Option #140
Option #141
Option #142
Option #143
Option #144
Option #145
Option #146
Option #147
Option #148
Option #149
Option #150
Option #151
Option #152
Option #153
Option #154
Option #155
Option #156
Option #157
Option #158
Option #159
Option #160
Option #161
Option #162
Option #163
Option #164
Option #165
Option #166
Option #167
Option #168
Option #169
Option #170
Option #171
Option #172
Option #173
Option #174
Option #175
Option #176
Option #177
Option #178
Option #179
Option #180
Option #181
Option #182
Option #183
Option #184
Option #185
Option #186
Option #187
Option #188
Option #189
Option #190
Option #191
Option #192
Option #193
Option #194
Option #195
Option #196
Option #197
Option #198
Option #199
Option #200
Option #201
Option #202
Option #203
Option #204
Option #205
Option #206
Option #207
Option #208
Option #209
Option #210
Option #211
Option #212
Option #213
Option #214
Option #215
Option #216
Option #217
Option #218
Option #219
Option #220
Option #221
Option #222
Option #223
Option #224
Option #225
Option #226
Option #227
Option #228
Option #229
Option #230
Option #231
Option #232
Option #233
Option #234
Option #235
Option #236
Option #237
Option #238
Option #239
Option #240
Option #241
Option #242
Option #243
Option #244
Option #245
Option #246
Option #247
Option #248
Option #249
Option #250
Option #251
Option #252
Option #253
Option #254
Option #255
Option #256
Option #257
Option #258
Option #259
Option #260
Option #261
Option #262
Option #263
Option #264
Option #265
Option #266
Option #267
Option #268
Option #269
Option #270
Option #271
Option #272
Option #273
Option #274
Option #275
Option #276
Option #277
Option #278
Option #279
Option #280
Option #281
Option #282
Option #283
Option #284
Option #285
Option #286
Option #287
Option #288
Option #289
Option #290
Option #291
Option #292
Option #293
Option #294
Option #295
Option #296
Option #297
Option #298
Option #299
Option #300
Option #301
Option #302
Option #303
Option #304
Option #305
Option #306
Option #307
Option #308
Option #309
Option #310
Option #311
Option #312
Option #313
Option #314
Option #315
Option #316
Option #317
Option #318
Option #319
Option #320
Option #321
Option #322
Option #323
Option #324
Option #325
Option #326
Option #327
Option #328
Option #329
Option #330
Option #331
Option #332
Option #333
Option #334
Option #335
Option #336
Option #337
Option #338
Option #339
Option #340
Option #341
Option #342
Option #343
Option #344
Option #345
Option #346
Option #347
Option #348
Option #349
Option #350
Option #351
Option #352
Option #353
Option #354
Option #355
Option #356
Option #357
Option #358
Option #359
Option #360
Option #361
Option #362
Option #363
Option #364
Option #365
Option #366
Option #367
Option #368
Option #369
Option #370
Option #371
Option #372
Option #373
Option #374
Option #375
Option #376
Option #377
Option #378
Option #379
Option #380
Option #381
Option #382
Option #383
Option #384
Option #385
Option #386
Option #387
Option #388
Option #389
Option #390
Option #391
Option #392
Option #393
Option #394
Option #395
Option #396
Option #397
Option #398
Option #399
Option #400
Option #401
Option #402
Option #403
Option #404
Option #405
Option #406
Option #407
Option #408
Option #409
Option #410
Option #411
Option #412
Option #413
Option #414
Option #415
Option #416
Option #417
Option #418
Option #419
Option #420
Option #421
Option #422
Option #423
Option #424
Option #425
Option #426
Option #427
Option #428
Option #429
Option #430
Option #431
Option #432
Option #433
Option #434
Option #435
Option #436
Option #437
Option #438
Option #439
Option #440
Option #441
Option #442
Option #443
Option #444
Option #445
Option #446
Option #447
Option #448
Option #449
Option #450
Option #451
Option #452
Option #453
Option #454
Option #455
Option #456
Option #457
Option #458
Option #459
Option #460
Option #461
Option #462
Option #463
Option #464
Option #465
Option #466
Option #467
Option #468
Option #469
Option #470
Option #471
Option #472
Option #473
Option #474
Option #475
Option #476
Option #477
Option #478
Option #479
Option #480
Option #481
Option #482
Option #483
Option #484
Option #485
Option #486
Option #487
Option #488
Option #489
Option #490
Option #491
Option #492
Option #493
Option #494
Option #495
Option #496
Option #497
Option #498
Option #499
Option #500
Option #501
Option #502
Option #503
Option #504
Option #505
Option #506
Option #507
Option #508
Option #509
Option #510
Option #511
Option #512
Option #513
Option #514
Option #515
Option #516
Option #517
Option #518
Option #519
Option #520
Option #521
Option #522
Option #523
Option #524
Option #525
Option #526
Option #527
Option #528
Option #529
Option #530
Option #531
Option #532
Option #533
Option #534
Option #535
Option #536
Option #537
Option #538
Option #539
Option #540
Option #541
Option #542
Option #543
Option #544
Option #545
Option #546
Option #547
Option #548
Option #549
Option #550
Option #551
Option #552
Option #553
Option #554
Option #555
Option #556
Option #557
Option #558
Option #559
Option #560
Option #561
Option #562
Option #563
Option #564
Option #565
Option #566
Option #567
Option #568
Option #569
Option #570
Option #571
Option #572
Option #573
Option #574
Option #575
Option #576
Option #577
Option #578
Option #579
Option #580
Option #581
Option #582
Option #583
Option #584
Option #585
Option #586
Option #587
Option #588
Option #589
Option #590
Option #591
Option #592
Option #593
Option #594
Option #595
Option #596
Option #597
Option #598
Option #599
Option #600
Option #601
Option #602
Option #603
Option #604
Option #605
Option #606
Option #607
Option #608
Option #609
Option #610
Option #611
Option #612
Option #613
Option #614
Option #615
Option #616
Option #617
Option #618
Option #619
Option #620
Option #621
Option #622
Option #623
Option #624
Option #625
Option #626
Option #627
Option #628
Option #629
Option #630
Option #631
Option #632
Option #633
Option #634
Option #635
Option #636
Option #637
Option #638
Option #639
Option #640
Option #641
Option #642
Option #643
Option #644
Option #645
Option #646
Option #647
Option #648
Option #649
Option #650
Option #651
Option #652
Option #653
Option #654
Option #655
Option #656
Option #657
Option #658
Option #659
Option #660
Option #661
Option #662
Option #663
Option #664
Option #665
Option #666
Option #667
Option #668
Option #669
Option #670
Option #671
Option #672
Option #673
Option #674
Option #675
Option #676
Option #677
Option #678
Option #679
Option #680
Option #681
Option #682
Option #683
Option #684
Option #685
Option #686
Option #687
Option #688
Option #689
Option #690
Option #691
Option #692
Option #693
Option #694
Option #695
Option #696
Option #697
Option #698
Option #699
Option #700
Option #701
Option #702
Option #703
Option #704
Option #705
Option #706
Option #707
Option #708
Option #709
Option #710
Option #711
Option #712
Option #713
Option #714
Option #715
Option #716
Option #717
Option #718
Option #719
Option #720
Option #721
Option #722
Option #723
Option #724
Option #725
Option #726
Option #727
Option #728
Option #729
Option #730
Option #731
Option #732
Option #733
Option #734
Option #735
Option #736
Option #737
Option #738
Option #739
Option #740
Option #741
Option #742
Option #743
Option #744
Option #745
Option #746
Option #747
Option #748
Option #749
Option #750
Option #751
Option #752
Option #753
Option #754
Option #755
Option #756
Option #757
Option #758
Option #759
Option #760
Option #761
Option #762
Option #763
Option #764
Option #765
Option #766
Option #767
Option #768
Option #769
Option #770
Option #771
Option #772
Option #773
Option #774
Option #775
Option #776
Option #777
Option #778
Option #779
Option #780
Option #781
Option #782
Option #783
Option #784
Option #785
Option #786
Option #787
Option #788
Option #789
Option #790
Option #791
Option #792
Option #793
Option #794
Option #795
Option #796
Option #797
Option #798
Option #799
Option #800
Option #801
Option #802
Option #803
Option #804
Option #805
Option #806
Option #807
Option #808
Option #809
Option #810
Option #811
Option #812
Option #813
Option #814
Option #815
Option #816
Option #817
Option #818
Option #819
Option #820
Option #821
Option #822
Option #823
Option #824
Option #825
Option #826
Option #827
Option #828
Option #829
Option #830
Option #831
Option #832
Option #833
Option #834
Option #835
Option #836
Option #837
Option #838
Option #839
Option #840
Option #841
Option #842
Option #843
Option #844
Option #845
Option #846
Option #847
Option #848
Option #849
Option #850
Option #851
Option #852
Option #853
Option #854
Option #855
Option #856
Option #857
Option #858
Option #859
Option #860
Option #861
Option #862
Option #863
Option #864
Option #865
Option #866
Option #867
Option #868
Option #869
Option #870
Option #871
Option #872
Option #873
Option #874
Option #875
Option #876
Option #877
Option #878
Option #879
Option #880
Option #881
Option #882
Option #883
Option #884
Option #885
Option #886
Option #887
Option #888
Option #889
Option #890
Option #891
Option #892
Option #893
Option #894
Option #895
Option #896
Option #897
Option #898
Option #899
Option #900
Option #901
Option #902
Option #903
Option #904
Option #905
Option #906
Option #907
Option #908
Option #909
Option #910
Option #911
Option #912
Option #913
Option #914
Option #915
Option #916
Option #917
Option #918
Option #919
Option #920
Option #921
Option #922
Option #923
Option #924
Option #925
Option #926
Option #927
Option #928
Option #929
Option #930
Option #931
Option #932
Option #933
Option #934
Option #935
Option #936
Option #937
Option #938
Option #939
Option #940
Option #941
Option #942
Option #943
Option #944
Option #945
Option #946
Option #947
Option #948
Option #949
Option #950
Option #951
Option #952
Option #953
Option #954
Option #955
Option #956
Option #957
Option #958
Option #959
Option #960
Option #961
Option #962
Option #963
Option #964
Option #965
Option #966
Option #967
Option #968
Option #969
Option #970
Option #971
Option #972
Option #973
Option #974
Option #975
Option #976
Option #977
Option #978
Option #979
Option #980
Option #981
Option #982
Option #983
Option #984
Option #985
Option #986
Option #987
Option #988
Option #989
Option #990
Option #991
Option #992
Option #993
Option #994
Option #995
Option #996
Option #997
Option #998
Option #999
Option #1000

PASS Behavior of Home and End for customizable-<select>
PASS Behavior of Home and End for customizable-<select>, starting at the top
PASS Behavior of Home and End for customizable-<select>, starting at the bottom
FAIL Behavior of PageUp and PageDown for customizable-<select> assert_equals: First page down should not scroll expected 305.390625 but got -240.609375
